// src/app/movies/models/tmdb-raw-response.ts
export interface TmdbRawMovieResponse {
  vote_count?: number;
  id: number;
  video?: boolean;
  vote_average?: number;
  title: string;
  popularity: number;
  poster_path: string;
  original_language?: string;
  original_title?: string;
  genre_ids?: number[];
  backdrop_path?: string;
  adult?: boolean;
  overview: string;
  release_date: Date;
}

export interface TmdbRawResponse {
  page: number;
  total_results: number;
  total_pages: number;
  results: TmdbRawMovieResponse[];
}